Clean up modules in actuation, extension and input (#1606)
