+++ module/actuation/FootController/data/config/webots/FootController.yaml
@@ -1,7 +1,3 @@
-log_level: INFO
-
-mode: IK # Options: IK
-
 startup:
   # Gain to use in startup phase
   servo_gain: 30
