--- module/actuation/Kinematics/data/config/nugus/Kinematics.yaml
@@ -1,16 +1,0 @@
-# Controls the minimum log level that NUClear log will display
-log_level: INFO
-
-urdf_path: "models/robot.urdf"
-
-# Stopping tolerance for IK
-ik_tolerance: 1e-5
-# Maximum number of iterations for IK
-ik_max_iterations: 500
-# IK method to use, either JACOBIAN, NLOPT, LEVENBERG_MARQUARDT, PARTICLE_SWARM or BFGS
-ik_method: LEVENBERG_MARQUARDT
-
-links:
-  torso: "torso"
-  left_foot: "left_foot_base"
-  right_foot: "right_foot_base"
