+++ module/actuation/KinematicsConfiguration/data/config/webots/KinematicsConfiguration.yaml
@@ -1,100 +1,4 @@
-leg:
-  hip_offset: [0.00, 0.055, 0.045] # [x, y, z]
-  upper_leg_length: 0.2
-  lower_leg_length: 0.2
-  heel_length: 0.085 # measured
-  foot_centre_to_ankle_centre: 0.02 # rough
-  foot:
-    width: 0.130 # rough
-    height: 0.04
-    length: 0.215 # rough
-    toe_length: 0.13 # measured
-  left_to_right:
-    hip_yaw: -1
-    hip_roll: -1
-    hip_pitch: 1
-    knee: 1
-    ankle_pitch: 1
-    ankle_roll: -1
 head:
-  camera_declination_angle_offset: pi / 90.0 # 2 degrees
-  neck_to_camera: [0.069, 0, 0.065] # [x, y, z]
-  ipd: 0.068 # Interpupillary distance - y axis
-  neck: # not yet calculated
-    length: 0.048
-    base_position_from_origin: [-0.007, 0, 0.21] # [x, y, z]
-  # head movement limits
   limits: # pitch limit is a function of yaw ... need to calculate this properly
     yaw: [-pi/2, pi/2] # [min_yaw, max_yaw]
     pitch: [-pi/2, pi/2] # [min_pitch, max_pitch]
-# rough measurements
-arm:
-  distance_between_shoulders: 0.17 # distance between shoulders
-  shoulder:
-    length: 0.0
-    width: 0.04
-    height: 0.010
-    offset: [0.0, 0.1905] # [x, z]
-  upper_arm:
-    length: 0.16
-    offset: [0, 0.03] # [y, x] very rough
-  lower_arm:
-    length: 0.235
-    offset: [0, 0] # [y, z] very rough
-mass_model:
-  # Particles CoM is in particle space
-  # CoM_x, CoM_y, CoM_z, mass
-  # SI units
-  particles:
-    head: [-0.05482228998067195, -0.00010888238321339966, 0.0011712900228573724, 0.41900000000000004]
-    arm_upper: [-0.0004754647829298364, 0.0004830631532648927, -0.047851449636613815, 0.29700000000000004]
-    arm_lower: [0.05336545685394963, -0.0008156196838209869, 0.0006063260979097202, 0.26]
-    torso: [0.0050079703254699804, 0.008880921777563762, 0.1021503146237477, 2.5390000000000006]
-    hip_block: [0.01573294687984765, -0.0005421722579394279, 0.0012617565679100488, 0.306]
-    leg_upper: [-0.0868239473635843, -0.0009904680702090176, -0.0032236573318317092, 0.387]
-    leg_lower: [-0.10000847255111779, 0.0012849442073489777, -0.0070729066189781946, 0.177]
-    ankle_block: [-0.01573294687984541, 0.0005126082201406932, -0.02197623716321553, 0.306]
-    foot: [0.030807480494529724, 0.008771710644510001, 0.011298451871418577, 0.20500000000000002]
-tensor_model:
-  # Particle inertial tensors (in partical space)
-  # [[xx, xy, xz],
-  #  [xy, yy, yz],
-  #  [xz, yz, zz]]
-  # SI units
-  particles:
-    head:
-      - [0.005229542632240126, 8.068689130285243e-07, -7.854193775142716e-05]
-      - [8.068689130285243e-07, 0.004273822060223504, 2.2633217085057325e-07]
-      - [-7.854193775142716e-05, 2.2633217085057325e-07, 0.0033336320526664368]
-    arm_upper:
-      - [0.002440902596230906, -3.6713691248137597e-06, -0.00013891452285131982]
-      - [-3.6713691248137597e-06, 0.002283834053695376, -1.9795333785260962e-05]
-      - [-0.00013891452285131982, -1.9795333785260962e-05, 0.0005366937214088004]
-    arm_lower:
-      - [0.0003407158555996035, 0.00011823673412134757, -2.8830788618180246e-05]
-      - [0.00011823673412134757, 0.0017692134503461817, 4.6856873014962955e-06]
-      - [-2.8830788618180246e-05, 4.6856873014962955e-06, 0.0016180592541080458]
-    torso:
-      - [0.01665728720713129, 0.0015451858906936002, 0.004938657742879685]
-      - [0.0015451858906936002, 0.004225235417508923, 0.0012027533896274989]
-      - [0.004938657742879685, 0.0012027533896274989, 0.05550778391062456]
-    hip_block:
-      - [0.0001179296979195992, 3.750777661643139e-06, -3.7623939617341825e-06]
-      - [3.750777661643139e-06, 0.00021675292165888234, -1.0471910235692515e-06]
-      - [-3.7623939617341825e-06, -1.0471910235692515e-06, 0.000261935711080282]
-    leg_upper:
-      - [0.0006967005196318071, 9.534768064209597e-06, -0.00047062464986191157]
-      - [9.534768064209597e-06, 0.0023927878790607625, -1.168766364373783e-07]
-      - [-0.00047062464986191157, -1.168766364373783e-07, 0.0020569215256885614]
-    leg_lower:
-      - [0.00026298531488117623, 8.38322526058631e-06, 0.0002634339989834785]
-      - [8.38322526058631e-06, 0.001587091949550628, 6.909200774158671e-07]
-      - [0.0002634339989834785, 6.909200774158671e-07, 0.001417720950622744]
-    ankle_block:
-      - [0.00011792969791959952, 3.7507776616440774e-06, 3.762393961734002e-06]
-      - [3.7507776616440774e-06, 0.00021675292165888193, 1.0471910235692212e-06]
-      - [3.762393961734002e-06, 1.0471910235692212e-06, 0.0002619357110802831]
-    foot:
-      - [0.00046922761556703155, -1.237289853891243e-05, 5.165617444725389e-06]
-      - [-1.237289853891243e-05, 0.0010836186025044235, 1.7594150186636352e-06]
-      - [5.165617444725389e-06, 1.7594150186636352e-06, 0.0013901177557110861]
